refactor: 작성페이지 상태 여부 체크하는 로직 layout -> stepMenuItem 으로 이동
